Основные фиксы нейминга. Фиксы работы потоков. Фиксы рефлексии. Выполнено из моего списка: 12. Сделать валидация главной старницы с вводом телефона 13. Сделать валидацию страницы с вводом реквест-кода 14. Сменить названия стран с официального на общее 15. Проверить респонс от сервера (ошибка на сервере) 16. Поправить текст на странице ввода реквест-кода 17. Автоматически посылать запрос проверки реквест кода на сервер при считывании кода из смс (Требуется детальная проверка)

+++ app/src/main/java/com/github/sasd97/upitter/models/response/coordinates/CoordinatesResponseModel.java
@@ -13,17 +13,17 @@ public class CoordinatesResponseModel extends BaseResponseModel {
 
     @SerializedName("latitude")
     @Expose
-    private float mLatitude;
+    private double mLatitude;
 
     @SerializedName("longitude")
     @Expose
-    private float mLongitude;
+    private double mLongitude;
 
-    public float getLatitude() {
+    public double getLatitude() {
         return mLatitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return mLongitude;
     }
